v1.4

+++ src/ilmasto.cpp
@@ -524,7 +524,7 @@ int main(void) {
 
 		if (mode.getValue() == "Manual") {
 			if (setup.getValue(0) != hertz) {
-				hertz = setup.getValue(0);
+				hertz = (int)setup.getValue(0);
 
 				setFrequency(node, (400 * hertz));
 				status.setValue(0, "OK");
